新增运行中插话拦截器：支持用户在任务执行时实时注入补充或纠偏指令到工作记忆，包含守卫体系防中断工具调用、首轮不注入保证初始意图、残留兜底及历史持久化

+++ soloncode-cli/src/main/java/org/noear/solon/codecli/portal/web/event/payload/SystemContextPayload.java
@@ -14,6 +14,6 @@ import java.io.Serializable;
 public class SystemContextPayload implements Serializable {
     private Integer tokens;
     private Integer count;
-    private Integer contextLimit;
+    private Long contextLimit;
     private Double cacheRate;
 }
